use Bun/global fetch, improve resources check, update lib entries
