+++ package.json
@@ -10,7 +10,7 @@
     "data:validate": "ajv validate -s react-native-libraries.schema.json -d react-native-libraries.json --verbose --allowUnionTypes",
     "libraries:cleanup": "bun scripts/cleanup-libraries-json && bun libraries:format",
     "libraries:recalculate": "bun scripts/recalculate-scores",
-    "libraries:format": "prettier --write react-native-libraries.json",
+    "libraries:format": "prettier --experimental-cli --write react-native-libraries.json",
     "libraries:check": "bun scripts/check-resources",
     "ci:cleanup-blobs": "bun scripts/cleanup-data-blobs",
     "ci:validate": "bun scripts/validate-new-entries",
@@ -25,7 +25,7 @@
     "@react-native-picker/picker": "^2.11.2",
     "@sentry/react": "^10.12.0",
     "@vercel/blob": "^0.27.3",
-    "expo": "54.0.8",
+    "expo": "54.0.9",
     "expo-font": "^14.0.8",
     "lodash": "^4.17.21",
     "next": "^15.5.3",
@@ -49,9 +49,8 @@
     "ajv-cli": "^5.0.0",
     "browserslist": "^4.26.2",
     "cheerio": "^1.1.2",
-    "cross-fetch": "^4.1.0",
     "dotenv": "^17.2.2",
-    "eslint": "^9.35.0",
+    "eslint": "^9.36.0",
     "eslint-config-next": "^15.5.3",
     "eslint-config-universe": "^15.0.3",
     "lint-staged": "^16.1.6",
