+++ scripts/helpers.ts
@@ -1,4 +1,4 @@
-import fetch from 'cross-fetch';
+import { fetch } from 'bun';
 
 import { LibraryType } from '~/types';
 
@@ -29,22 +29,6 @@ export async function getUpdatedUrl(url: string) {
   } catch {
     return url;
   }
-}
-
-export function parseGitHubUrl(url: string) {
-  const [, , , repoOwner, repoName, ...path] = url.split('/');
-
-  const isMonorepo = !!(path && path.length);
-  const branchName = path[1];
-  const packagePath = isMonorepo ? path.slice(2).join('/').replace('%40', '@') : '.';
-
-  return {
-    repoOwner,
-    repoName,
-    isMonorepo,
-    branchName,
-    packagePath,
-  };
 }
 
 export function sleep(ms = 0, msMax: number | undefined = undefined) {
